services/ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
redis==5.0.1
celery==5.3.4
httpx==0.25.2
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-exporter-jaeger==1.21.0
structlog==23.2.0
python-json-logger==2.0.7

# AI/ML Dependencies
torch==2.1.1
torchvision==0.16.1
transformers==4.36.0
sentence-transformers==2.2.2
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.4
pandas==2.1.4
scikit-learn==1.3.2
joblib==1.3.2
openai==1.3.7
anthropic==0.7.8
langchain==0.0.352
langchain-openai==0.0.2
langchain-anthropic==0.0.1
pydantic-ai==0.0.12

# Computer Vision
easyocr==1.7.0
pytesseract==0.3.10

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0

# Model Monitoring
evidently==0.4.11
mlflow==2.8.1 